+++ flutter_app/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'login.dart';
+import 'search_result.dart';
 
 void main() {
   runApp(const MyApp());
@@ -127,8 +128,6 @@ class _SearchScreenState extends State<SearchScreen> {
   final _hobbiesController = TextEditingController();
   final List<String> _ageGroups = ['20代', '30代', '40代', '50代', '60代以上'];
   String? _selectedAgeGroup;
-  List<Map<String, String>> _searchResults = [];
-  bool _isLoading = false;
   String? _errorMessage;
 
   // 検索処理
@@ -137,55 +136,20 @@ class _SearchScreenState extends State<SearchScreen> {
     if (_selectedAgeGroup == null || _hobbiesController.text.trim().isEmpty) {
       setState(() {
         _errorMessage = '趣味と年代の両方を入力してください。';
-        _searchResults = []; // エラー時は検索結果をクリア
       });
       return;
     }
 
-    setState(() {
-      _isLoading = true;
-      _errorMessage = null;
-      _searchResults = [];
-    });
-
-    // モックデータで検索結果をシミュレート
-    Future.delayed(const Duration(seconds: 1), () {
-      setState(() {
-        _searchResults = [
-          {
-            'name': '山田 太郎',
-            'ageGroup': _selectedAgeGroup!,
-            'hobbies': _hobbiesController.text,
-            'department': '開発部',
-          },
-          {
-            'name': '佐藤 花子',
-            'ageGroup': _selectedAgeGroup!,
-            'hobbies': _hobbiesController.text,
-            'department': '企画部',
-          },
-          {
-            'name': '鈴木 一郎',
-            'ageGroup': _selectedAgeGroup!,
-            'hobbies': _hobbiesController.text,
-            'department': '営業部',
-          },
-          {
-            'name': '田中 美咲',
-            'ageGroup': _selectedAgeGroup!,
-            'hobbies': _hobbiesController.text,
-            'department': '人事部',
-          },
-          {
-            'name': '伊藤 健太',
-            'ageGroup': _selectedAgeGroup!,
-            'hobbies': _hobbiesController.text,
-            'department': '経理部',
-          },
-        ];
-        _isLoading = false;
-      });
-    });
+    // 検索結果画面に遷移
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (context) => SearchResultScreen(
+          hobbies: _hobbiesController.text.trim(),
+          ageGroup: _selectedAgeGroup!,
+        ),
+      ),
+    );
   }
 
   @override
@@ -231,274 +195,90 @@ class _SearchScreenState extends State<SearchScreen> {
           child: Container(color: Colors.grey.shade300, height: 1.0),
         ),
       ),
-      body: Center(
-        child: ConstrainedBox(
-          constraints: const BoxConstraints(maxWidth: 1200),
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 24.0),
-            child: Row(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                // --- 検索フォーム部分 ---
-                Expanded(
-                  flex: 1,
-                  child: Container(
-                    margin: const EdgeInsets.only(top: 32.0, bottom: 32.0, right: 16.0),
-                    padding: const EdgeInsets.all(32.0),
+      body: Container(
+        decoration: BoxDecoration(
+          gradient: LinearGradient(
+            begin: Alignment.topCenter,
+            end: Alignment.bottomCenter,
+            colors: [
+              const Color(0xFFFAFAFA),
+              Colors.grey.shade50,
+            ],
+          ),
+        ),
+        child: Center(
+          child: ConstrainedBox(
+            constraints: const BoxConstraints(maxWidth: 600),
+            child: Padding(
+              padding: const EdgeInsets.all(24.0),
+              child: Column(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  // メインの検索フォーム
+                  Container(
+                    padding: const EdgeInsets.all(40.0),
                     decoration: BoxDecoration(
                       color: Colors.white,
-                      borderRadius: BorderRadius.circular(20.0),
+                      borderRadius: BorderRadius.circular(24.0),
                       border: Border.all(color: Colors.grey.shade100, width: 1.0),
                       boxShadow: [
                         BoxShadow(
-                          color: Colors.black.withOpacity(0.06),
-                          blurRadius: 20,
-                          offset: const Offset(0, 4),
-                        ),
-                      ],
-                    ),
-                    child: SingleChildScrollView(
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Row(
-                            children: [
-                              Container(
-                                padding: const EdgeInsets.all(12),
-                                decoration: BoxDecoration(
-                                  gradient: LinearGradient(
-                                    colors: [
-                                      const Color(0xFF6366F1),
-                                      const Color(0xFF8B5CF6),
-                                    ],
-                                    begin: Alignment.topLeft,
-                                    end: Alignment.bottomRight,
-                                  ),
-                                  borderRadius: BorderRadius.circular(12),
-                                ),
-                                child: const Icon(
-                                  Icons.search,
-                                  color: Colors.white,
-                                  size: 20,
-                                ),
-                              ),
-                              const SizedBox(width: 16),
-                              Expanded(
-                                child: Column(
-                                  crossAxisAlignment: CrossAxisAlignment.start,
-                                  children: [
-                                    const Text(
-                                      '検索条件',
-                                      style: TextStyle(
-                                        fontSize: 24,
-                                        fontWeight: FontWeight.bold,
-                                        color: Color(0xFF111827),
-                                      ),
-                                    ),
-                                    const SizedBox(height: 4),
-                                    Text(
-                                      'あなたの趣味や年代に合う人を探してみましょう',
-                                      style: TextStyle(
-                                        fontSize: 15,
-                                        color: Colors.grey.shade600,
-                                        height: 1.4,
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                              ),
-                            ],
-                          ),
-                          const SizedBox(height: 40),
-
-                          const Text(
-                            '趣味',
-                            style: TextStyle(
-                              fontSize: 16,
-                              fontWeight: FontWeight.bold,
-                              color: Color(0xFF374151),
-                            ),
-                          ),
-                          const SizedBox(height: 12),
-                          TextField(
-                            controller: _hobbiesController,
-                            decoration: const InputDecoration(
-                              hintText: '例: 読書, 旅行, 映画鑑賞',
-                              prefixIcon: Icon(
-                                Icons.favorite_outline,
-                                color: Color(0xFF6366F1),
-                                size: 20,
-                              ),
-                            ),
-                          ),
-                          const SizedBox(height: 32),
-
-                          const Text(
-                            '年代',
-                            style: TextStyle(
-                              fontSize: 16,
-                              fontWeight: FontWeight.bold,
-                              color: Color(0xFF374151),
-                            ),
-                          ),
-                          const SizedBox(height: 12),
-                          DropdownButtonFormField<String>(
-                            value: _selectedAgeGroup,
-                            decoration: InputDecoration(
-                              hintText: '年代を選択してください',
-                              prefixIcon: const Icon(
-                                Icons.calendar_today,
-                                color: Color(0xFF6366F1),
-                                size: 20,
-                              ),
-                              contentPadding: const EdgeInsets.symmetric(
-                                horizontal: 16,
-                                vertical: 16,
-                              ),
-                              filled: true,
-                              fillColor: Colors.white,
-                              border: OutlineInputBorder(
-                                borderRadius: BorderRadius.circular(12.0),
-                                borderSide: BorderSide(
-                                  color: Colors.grey.shade200,
-                                  width: 1.0,
-                                ),
-                              ),
-                            ),
-                            style: const TextStyle(
-                              fontSize: 16,
-                              color: Color(0xFF374151),
-                            ),
-                            icon: const Icon(
-                              Icons.arrow_drop_down,
-                              color: Color(0xFF6366F1),
-                            ),
-                            isExpanded: true,
-                            hint: Text(
-                              '年代を選択してください',
-                              style: TextStyle(color: Colors.grey.shade500),
-                            ),
-                            items: _ageGroups.map<DropdownMenuItem<String>>((
-                              String value,
-                            ) {
-                              return DropdownMenuItem<String>(
-                                value: value,
-                                child: Text(value),
-                              );
-                            }).toList(),
-                            onChanged: (String? newValue) =>
-                                setState(() => _selectedAgeGroup = newValue),
-                          ),
-                          const SizedBox(height: 48),
-
-                          SizedBox(
-                            width: double.infinity,
-                            height: 56,
-                            child: ElevatedButton(
-                              onPressed: _isLoading ? null : _search,
-                              style: ElevatedButton.styleFrom(
-                                backgroundColor: _isLoading 
-                                  ? Colors.grey.shade300 
-                                  : const Color(0xFF6366F1),
-                                elevation: 0,
-                                shadowColor: Colors.transparent,
-                                shape: RoundedRectangleBorder(
-                                  borderRadius: BorderRadius.circular(16.0),
-                                ),
-                              ),
-                              child: _isLoading
-                                  ? const SizedBox(
-                                      width: 20,
-                                      height: 20,
-                                      child: CircularProgressIndicator(
-                                        strokeWidth: 2,
-                                        valueColor: AlwaysStoppedAnimation<Color>(
-                                          Colors.white,
-                                        ),
-                                      ),
-                                    )
-                                  : const Row(
-                                      mainAxisAlignment: MainAxisAlignment.center,
-                                      children: [
-                                        Icon(Icons.search, size: 20),
-                                        SizedBox(width: 8),
-                                        Text(
-                                          '検索する',
-                                          style: TextStyle(
-                                            fontSize: 16,
-                                            fontWeight: FontWeight.w600,
-                                          ),
-                                        ),
-                                      ],
-                                    ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ),
-                ),
-
-                // --- 検索結果表示部分 ---
-                Expanded(
-                  flex: 2,
-                  child: Container(
-                    margin: const EdgeInsets.only(top: 32.0, bottom: 32.0, left: 16.0),
-                    padding: const EdgeInsets.all(32.0),
-                    decoration: BoxDecoration(
-                      color: Colors.white,
-                      borderRadius: BorderRadius.circular(20.0),
-                      border: Border.all(color: Colors.grey.shade100, width: 1.0),
-                      boxShadow: [
-                        BoxShadow(
-                          color: Colors.black.withOpacity(0.06),
-                          blurRadius: 20,
-                          offset: const Offset(0, 4),
+                          color: Colors.black.withOpacity(0.08),
+                          blurRadius: 24,
+                          offset: const Offset(0, 8),
                         ),
                       ],
                     ),
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: [
+                        // ヘッダー部分
                         Row(
                           children: [
                             Container(
-                              padding: const EdgeInsets.all(12),
+                              padding: const EdgeInsets.all(16),
                               decoration: BoxDecoration(
                                 gradient: LinearGradient(
                                   colors: [
-                                    const Color(0xFF10B981),
-                                    const Color(0xFF059669),
+                                    const Color(0xFF6366F1),
+                                    const Color(0xFF8B5CF6),
                                   ],
                                   begin: Alignment.topLeft,
                                   end: Alignment.bottomRight,
                                 ),
-                                borderRadius: BorderRadius.circular(12),
+                                borderRadius: BorderRadius.circular(16),
+                                boxShadow: [
+                                  BoxShadow(
+                                    color: const Color(0xFF6366F1).withOpacity(0.3),
+                                    blurRadius: 12,
+                                    offset: const Offset(0, 4),
+                                  ),
+                                ],
                               ),
                               child: const Icon(
-                                Icons.people,
+                                Icons.search,
                                 color: Colors.white,
-                                size: 20,
+                                size: 24,
                               ),
                             ),
-                            const SizedBox(width: 16),
+                            const SizedBox(width: 20),
                             const Expanded(
                               child: Column(
                                 crossAxisAlignment: CrossAxisAlignment.start,
                                 children: [
                                   Text(
-                                    '検索結果',
+                                    'プロフィール検索',
                                     style: TextStyle(
-                                      fontSize: 24,
+                                      fontSize: 28,
                                       fontWeight: FontWeight.bold,
                                       color: Color(0xFF111827),
                                     ),
                                   ),
                                   SizedBox(height: 4),
                                   Text(
-                                    '条件に合致する人が見つかりました',
+                                    'あなたの趣味や年代に合う人を探してみましょう',
                                     style: TextStyle(
-                                      fontSize: 15,
+                                      fontSize: 16,
                                       color: Colors.grey,
                                       height: 1.4,
                                     ),
@@ -508,13 +288,154 @@ class _SearchScreenState extends State<SearchScreen> {
                             ),
                           ],
                         ),
-                        const SizedBox(height: 32),
-                        Expanded(child: _buildResultView()),
+                        const SizedBox(height: 40),
+
+                        // エラーメッセージ表示
+                        if (_errorMessage != null) ...[
+                          Container(
+                            padding: const EdgeInsets.all(16),
+                            decoration: BoxDecoration(
+                              color: Colors.red.shade50,
+                              borderRadius: BorderRadius.circular(12),
+                              border: Border.all(color: Colors.red.shade200, width: 1),
+                            ),
+                            child: Row(
+                              children: [
+                                Icon(
+                                  Icons.error_outline,
+                                  color: Colors.red.shade600,
+                                  size: 20,
+                                ),
+                                const SizedBox(width: 12),
+                                Expanded(
+                                  child: Text(
+                                    _errorMessage!,
+                                    style: TextStyle(
+                                      color: Colors.red.shade700,
+                                      fontSize: 14,
+                                      fontWeight: FontWeight.w500,
+                                    ),
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
+                          const SizedBox(height: 24),
+                        ],
+
+                        // 趣味入力フィールド
+                        const Text(
+                          '趣味',
+                          style: TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.bold,
+                            color: Color(0xFF374151),
+                          ),
+                        ),
+                        const SizedBox(height: 12),
+                        TextField(
+                          controller: _hobbiesController,
+                          decoration: const InputDecoration(
+                            hintText: '例: 読書, 旅行, 映画鑑賞',
+                            prefixIcon: Icon(
+                              Icons.favorite_outline,
+                              color: Color(0xFF6366F1),
+                              size: 20,
+                            ),
+                          ),
+                          onChanged: (value) {
+                            if (_errorMessage != null) {
+                              setState(() {
+                                _errorMessage = null;
+                              });
+                            }
+                          },
+                        ),
+                        const SizedBox(height: 28),
+
+                        // 年代選択フィールド
+                        const Text(
+                          '年代',
+                          style: TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.bold,
+                            color: Color(0xFF374151),
+                          ),
+                        ),
+                        const SizedBox(height: 12),
+                        DropdownButtonFormField<String>(
+                          value: _selectedAgeGroup,
+                          decoration: const InputDecoration(
+                            hintText: '年代を選択してください',
+                            prefixIcon: Icon(
+                              Icons.calendar_today,
+                              color: Color(0xFF6366F1),
+                              size: 20,
+                            ),
+                          ),
+                          style: const TextStyle(
+                            fontSize: 16,
+                            color: Color(0xFF374151),
+                          ),
+                          icon: const Icon(
+                            Icons.arrow_drop_down,
+                            color: Color(0xFF6366F1),
+                          ),
+                          isExpanded: true,
+                          items: _ageGroups.map<DropdownMenuItem<String>>((
+                            String value,
+                          ) {
+                            return DropdownMenuItem<String>(
+                              value: value,
+                              child: Text(value),
+                            );
+                          }).toList(),
+                          onChanged: (String? newValue) {
+                            setState(() {
+                              _selectedAgeGroup = newValue;
+                              if (_errorMessage != null) {
+                                _errorMessage = null;
+                              }
+                            });
+                          },
+                        ),
+                        const SizedBox(height: 48),
+
+                        // 検索ボタン
+                        SizedBox(
+                          width: double.infinity,
+                          height: 56,
+                          child: ElevatedButton(
+                            onPressed: _search,
+                            style: ElevatedButton.styleFrom(
+                              backgroundColor: const Color(0xFF6366F1),
+                              elevation: 0,
+                              shadowColor: Colors.transparent,
+                              shape: RoundedRectangleBorder(
+                                borderRadius: BorderRadius.circular(16.0),
+                              ),
+                            ),
+                            child: const Row(
+                              mainAxisAlignment: MainAxisAlignment.center,
+                              children: [
+                                Icon(Icons.search, size: 20),
+                                SizedBox(width: 8),
+                                Text(
+                                  '検索する',
+                                  style: TextStyle(
+                                    fontSize: 16,
+                                    fontWeight: FontWeight.w600,
+                                  ),
+                                ),
+                              ],
+                            ),
+                          ),
+                        ),
                       ],
                     ),
                   ),
-                ),
-              ],
+                ],
+              ),
             ),
           ),
         ),
@@ -592,297 +513,6 @@ class _SearchScreenState extends State<SearchScreen> {
     );
   }
 
-  // --- ▼▼▼ 結果表示部分のロジックを分離して見やすくする ▼▼▼ ---
-  Widget _buildResultView() {
-    if (_isLoading) {
-      return const Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            CircularProgressIndicator(color: Color(0xFF6366F1)),
-            SizedBox(height: 16),
-            Text(
-              '検索中...',
-              style: TextStyle(
-                fontSize: 16,
-                color: Color(0xFF6366F1),
-                fontWeight: FontWeight.w500,
-              ),
-            ),
-          ],
-        ),
-      );
-    }
-    if (_errorMessage != null) {
-      return Center(
-        child: Container(
-          padding: const EdgeInsets.all(24),
-          decoration: BoxDecoration(
-            color: Colors.red.shade50,
-            borderRadius: BorderRadius.circular(16),
-            border: Border.all(color: Colors.red.shade200, width: 1),
-          ),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Icon(
-                Icons.error_outline,
-                color: Colors.red.shade600,
-                size: 48,
-              ),
-              const SizedBox(height: 16),
-              Text(
-                _errorMessage!,
-                style: TextStyle(
-                  color: Colors.red.shade700,
-                  fontSize: 16,
-                  fontWeight: FontWeight.w500,
-                ),
-                textAlign: TextAlign.center,
-              ),
-            ],
-          ),
-        ),
-      );
-    }
-    if (_searchResults.isEmpty) {
-      return Center(
-        child: Container(
-          padding: const EdgeInsets.all(24),
-          decoration: BoxDecoration(
-            color: Colors.grey.shade50,
-            borderRadius: BorderRadius.circular(16),
-            border: Border.all(color: Colors.grey.shade200, width: 1),
-          ),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              Icon(
-                Icons.search_off,
-                color: Colors.grey.shade500,
-                size: 48,
-              ),
-              const SizedBox(height: 16),
-              Text(
-                '条件に合う結果が見つかりませんでした',
-                style: TextStyle(
-                  color: Colors.grey.shade700,
-                  fontSize: 16,
-                  fontWeight: FontWeight.w500,
-                ),
-                textAlign: TextAlign.center,
-              ),
-              const SizedBox(height: 8),
-              Text(
-                '検索条件を変更してお試しください',
-                style: TextStyle(
-                  color: Colors.grey.shade600,
-                  fontSize: 14,
-                ),
-                textAlign: TextAlign.center,
-              ),
-            ],
-          ),
-        ),
-      );
-    }
-    return ListView.separated(
-      itemCount: _searchResults.length,
-      separatorBuilder: (context, index) => const SizedBox(height: 20),
-      itemBuilder: (context, index) {
-        final person = _searchResults[index];
-        return Container(
-          padding: const EdgeInsets.all(24.0),
-          decoration: BoxDecoration(
-            color: const Color(0xFFFAFAFA),
-            borderRadius: BorderRadius.circular(18.0),
-            border: Border.all(color: Colors.grey.shade200, width: 1.0),
-            boxShadow: [
-              BoxShadow(
-                color: Colors.black.withOpacity(0.08),
-                blurRadius: 12,
-                offset: const Offset(0, 4),
-              ),
-            ],
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                children: [
-                  Container(
-                    width: 56,
-                    height: 56,
-                    decoration: BoxDecoration(
-                      gradient: LinearGradient(
-                        colors: [
-                          const Color(0xFF6366F1),
-                          const Color(0xFF8B5CF6),
-                        ],
-                        begin: Alignment.topLeft,
-                        end: Alignment.bottomRight,
-                      ),
-                      borderRadius: BorderRadius.circular(28),
-                      boxShadow: [
-                        BoxShadow(
-                          color: const Color(0xFF6366F1).withOpacity(0.3),
-                          blurRadius: 8,
-                          offset: const Offset(0, 4),
-                        ),
-                      ],
-                    ),
-                    child: const Icon(
-                      Icons.person,
-                      color: Colors.white,
-                      size: 28,
-                    ),
-                  ),
-                  const SizedBox(width: 20),
-                  Expanded(
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          person['name'] ?? '名前なし',
-                          style: const TextStyle(
-                            fontSize: 20,
-                            fontWeight: FontWeight.bold,
-                            color: Color(0xFF111827),
-                          ),
-                        ),
-                        const SizedBox(height: 8),
-                        Container(
-                          padding: const EdgeInsets.symmetric(
-                            horizontal: 14,
-                            vertical: 8,
-                          ),
-                          decoration: BoxDecoration(
-                            gradient: LinearGradient(
-                              colors: [
-                                const Color(0xFF10B981).withOpacity(0.15),
-                                const Color(0xFF059669).withOpacity(0.15),
-                              ],
-                              begin: Alignment.centerLeft,
-                              end: Alignment.centerRight,
-                            ),
-                            borderRadius: BorderRadius.circular(10),
-                            border: Border.all(
-                              color: const Color(0xFF10B981).withOpacity(0.3),
-                              width: 1,
-                            ),
-                          ),
-                          child: Text(
-                            person['department'] ?? '部署不明',
-                            style: const TextStyle(
-                              fontSize: 14,
-                              color: Color(0xFF059669),
-                              fontWeight: FontWeight.w600,
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 24),
-              Container(
-                padding: const EdgeInsets.all(20),
-                decoration: BoxDecoration(
-                  color: Colors.white,
-                  borderRadius: BorderRadius.circular(16),
-                  border: Border.all(color: Colors.grey.shade200, width: 1),
-                  boxShadow: [
-                    BoxShadow(
-                      color: Colors.black.withOpacity(0.04),
-                      blurRadius: 8,
-                      offset: const Offset(0, 2),
-                    ),
-                  ],
-                ),
-                child: Column(
-                  children: [
-                    Row(
-                      children: [
-                        Container(
-                          padding: const EdgeInsets.all(8),
-                          decoration: BoxDecoration(
-                            color: const Color(0xFF6366F1).withOpacity(0.1),
-                            borderRadius: BorderRadius.circular(8),
-                          ),
-                          child: const Icon(
-                            Icons.calendar_today,
-                            size: 18,
-                            color: Color(0xFF6366F1),
-                          ),
-                        ),
-                        const SizedBox(width: 12),
-                        Text(
-                          '年代:',
-                          style: TextStyle(
-                            fontSize: 15,
-                            fontWeight: FontWeight.w600,
-                            color: Colors.grey.shade700,
-                          ),
-                        ),
-                        const SizedBox(width: 8),
-                        Text(
-                          person['ageGroup'] ?? '不明',
-                          style: const TextStyle(
-                            fontSize: 15,
-                            fontWeight: FontWeight.bold,
-                            color: Color(0xFF374151),
-                          ),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 16),
-                    Row(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Container(
-                          padding: const EdgeInsets.all(8),
-                          decoration: BoxDecoration(
-                            color: const Color(0xFFEC4899).withOpacity(0.1),
-                            borderRadius: BorderRadius.circular(8),
-                          ),
-                          child: const Icon(
-                            Icons.favorite,
-                            size: 18,
-                            color: Color(0xFFEC4899),
-                          ),
-                        ),
-                        const SizedBox(width: 12),
-                        Text(
-                          '趣味:',
-                          style: TextStyle(
-                            fontSize: 15,
-                            fontWeight: FontWeight.w600,
-                            color: Colors.grey.shade700,
-                          ),
-                        ),
-                        const SizedBox(width: 8),
-                        Expanded(
-                          child: Text(
-                            person['hobbies'] ?? '不明',
-                            style: const TextStyle(
-                              fontSize: 15,
-                              fontWeight: FontWeight.bold,
-                              color: Color(0xFF374151),
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        );
-      },
-    );
-  }
 
   @override
   void dispose() {
